move SecretField to local directory
