+++ frontend/pages/UserSettingsPage/APITokenModal/TokenSecretField/_styles.scss
